% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aineistoA_dat}
\alias{aineistoA_dat}
\title{aineistoA_dat}
\format{
A data frame with 50 rows and 9 variables:
\describe{
  \item{\code{lomakenro}}{integer Identifikaationa vastauslomakkeen numero}
  \item{\code{sukupuoli}}{integer Opiskelijan sukupuoli: 1=mies, 2=nainen}
  \item{\code{ikä}}{integer Opiskelijan ikä vuosina}
  \item{\code{tiedekunta}}{integer Opiskelijan tiedekunta: 1=yhteiskuntatieteiden, 2=kasvatustieteiden, 3=humanististen tieteiden}
  \item{\code{opiskeluvuosi}}{integer Opiskelijan tiedekunnassa meneillään oleva opiskeluvuosi}
  \item{\code{matem_pisteet}}{integer Ennen kurssia suoritetun matematiikan testin pistemäärä}
  \item{\code{tarpeellisuus}}{integer Vastaus kysymykseen "Kuinka tarpeellinen mielestäsi pakollinen tilastotieteen kurssi on omassa tutkinnossasi?": 1=täysin tarpeeton, 2=yhdentekevä, 3=hyvin tarpeellinen}
  \item{\code{välikoe1}}{integer Tilastotieteen ensimmäisen välikokeen pistemäärä}
  \item{\code{välikoe2}}{integer Tilastotieteen toisen välikokeen pistemäärä} 
}
}
\usage{
aineistoA_dat
}
\description{
Harjoitustehtävissä käytettävä aineistoA löytyy tästä.
}
\details{
DETAILS
}
\keyword{datasets}
