% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alko_tupakka}
\alias{alko_tupakka}
\title{Alkoholin ja tupakkatuotteiden kulutus}
\format{
A data frame with 11 rows and 3 variables:
\describe{
  \item{\code{alue}}{character Alue}
  \item{\code{alkoholi}}{double Alkoholin kulutus}
  \item{\code{tupakkatuotteet}}{double Tupakkatuotteiden kulutus}
}
}
\usage{
alko_tupakka
}
\description{
Esimerkkiaineisto alkoholin ja tupakkatuotteiden kulutuksesta Ison-Britannian alueilla.
}
\keyword{datasets}
