% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hlotsim_dat}
\alias{hlotsim_dat}
\title{hlotsim_dat}
\format{
A data frame with 40 rows and 5 variables:
\describe{
  \item{\code{sukupuoli}}{integer Sukupuoli}
  \item{\code{ikä}}{double Ikä vuosina}
  \item{\code{pituus}}{double Henkilön pituus cm}
  \item{\code{paino}}{double Henkilö paino kg}
  \item{\code{pääaine}}{integer Henkilön pääaine yliopistossa} 
}
}
\usage{
hlotsim_dat
}
\description{
Keinotekoinen aineisto luentoesimerkkiä varten
}
\details{
DETAILS
}
\keyword{datasets}
