% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{koneet_dat}
\alias{koneet_dat}
\title{Kahden koneen valmistamien mitallien painot}
\format{
A data frame with 16 rows and 2 variables:
\describe{
  \item{\code{mitalinpaino}}{double Valmistetun kappaleen paino}
  \item{\code{kone}}{factor Kone A tai Kone B}
}
}
\usage{
koneet_dat
}
\description{
Harjoitusaineisto kahden koneen valmistamien kappaleiden painojen vertailuun.
}
\keyword{datasets}
