% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kysely_dat}
\alias{kysely_dat}
\title{kysely.csv data}
\format{
A data frame with 186 rows and 2 variables:
\describe{
  \item{\code{sukupuoli}}{integer Kyselyyn vastanneen sukupuoli}
  \item{\code{itseopiskelu}}{double Vastaus kysymykseen: “Luentosarjaan kuuluu 4 viikkotuntia luentoja. Kuinka monta tuntia viikossa arvioit käyttäväsi tämän lisäksi kurssin itsenäiseen opiskeluun keskimäärin?”} 
}
}
\usage{
kysely_dat
}
\description{
Aineisto kysely_dat vastaa kysely.csv -tiedoston aineistoa.
}
\details{
DETAILS
}
\keyword{datasets}
