% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcb.data}
\alias{pcb.data}
\title{PCB-pitoisuudet vuosina 2018 ja 2020}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{\code{aika}}{factor Mittausvuosi: 2018 tai 2020}
  \item{\code{pcb}}{double PCB-pitoisuus}
}
}
\usage{
pcb.data
}
\description{
Esimerkkiaineisto kahden ajankohdan PCB-pitoisuuksista.
}
\keyword{datasets}
