% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viini_dat}
\alias{viini_dat}
\title{Viininkulutus ja sydäntautikuolleisuus}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{\code{country}}{character Maa}
  \item{\code{alko}}{double Vuotuinen viininkulutus}
  \item{\code{death}}{double Sydäntautikuolleisuus}
}
}
\usage{
viini_dat
}
\description{
Esimerkkiaineisto viininkulutuksen ja sydäntautikuolleisuuden yhteyden tarkasteluun.
}
\keyword{datasets}
